+++ packages/builder-vue2/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @varlinor/builder-vue2
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @varlinor/node-tools@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @varlinor/node-tools@2.0.1
+
 ## 2.0.0
 
 ### Major Changes