+++ packages/node-tools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @varlinor/node-tools
 
+## 2.0.2
+
+### Patch Changes
+
+- feat: support to create hook for unbuild and plugin for vite.
+
+## 2.0.1
+
+### Patch Changes
+
+- fix: support auto scan packages info and auto fix dir import among multiple packages
+
 ## 2.0.0
 
 ### Major Changes